% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{data}
\name{veteranos}
\alias{veteranos}
\title{Veteranos.}
\format{
Un data frame con 7 columnas:
\describe{
\item{\code{Apellidos}}{texto}
\item{\code{Nombres}}{texto}
\item{\code{Arma}}{texto}
\item{\code{Grado}}{texto}
\item{\code{Categoria}}{texto}
\item{\code{Sexo}}{
  texto. Nota: Derivado del listado "Veteranas", no estoy seguro si refiere al sexo o al género,
  debo averiguar.
}
\item{\code{Documento}}{numérico}
}
}
\usage{
veteranos
}
\description{
Nómina de Veteranos de Guerra de Malvinas.
}
\keyword{datasets}
